--- Lv.0/Soltuion120805.java
@@ -1,5 +1,0 @@
-class Soltuion120805 {
-    public int solution(int num1, int num2) {
-        return num1 / num2;
-    }
-}
