--- Lv.0/Solution120803.java
@@ -1,5 +1,0 @@
-class Solution120803 {
-    public int solution(int num1, int num2) {
-        return num1 - num2;
-    }
-}
