--- Lv.0/Solution120804.java
@@ -1,5 +1,0 @@
-class Solution120804 {
-    public int solution(int num1, int num2) {
-        return num1 * num2;
-    }
-}
