--- Lv.0/Solution120810.java
@@ -1,5 +1,0 @@
-class Solution120810 {
-    public int soultion(int num1, int num2) {
-        return num1 % num2;
-    }
-}
